Refactor tensorflow tests
