+++ tests/unit/tensorflow/utils/test_io.py
