+++ tests/unit/tensorflow/utils/test_ops.py
