+++ tests/unit/tensorflow/utils/test_plotting.py
