+++ tests/unit/tensorflow/utils/test_settings.py
